customer changes

+++ src/main/java/br/com/brasilprev/api/service/CustomerService.java
@@ -1,17 +1,23 @@
 package br.com.brasilprev.api.service;
 
+import br.com.brasilprev.api.model.Order;
 import br.com.brasilprev.api.model.dto.CustomerDTO;
 import br.com.brasilprev.api.model.Customer;
 import br.com.brasilprev.api.model.mapper.CustomerMapper;
 import br.com.brasilprev.api.repository.CustomerRepository;
+import br.com.brasilprev.api.repository.OrderRepository;
+import com.sun.org.apache.bcel.internal.generic.ATHROW;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,6 +28,9 @@ public class CustomerService {
 
     @Autowired
     private CustomerMapper customerMapper;
+
+    @Autowired
+    private OrderRepository orderRepository;
 
     @Autowired
     private MessageSource messageSource;
@@ -51,6 +60,11 @@ public class CustomerService {
     }
 
     public void delete(Long id) {
+        Set<Order> orderSet = new HashSet<>(orderRepository.findAll());
+        if(orderSet.stream().anyMatch(order -> order.getCustomer().getId().equals(id))){
+            throw new DataIntegrityViolationException(
+                    messageSource.getMessage("resource.operation-not-allowed-explained0", null, LocaleContextHolder.getLocale()));
+        }
         customerRepository.delete(id);
     }
